src/deploy/requirements.txt
azure-identity
azure-mgmt-resource
azure-mgmt-web
azure-storage-blob
python-dotenv
